data-analyst-agent-agent-api-20250715-170220-7803907e-main/teams/requirements.txt
agno>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
# Team Studio requirements with enhanced features
openai>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
loguru>=0.7.0